requirement.txt
PyJWT==2.8.0
redis==5.0.1
requests==2.31.0
